requirements.txt
google-genai
networkx
pyvis
pyyaml
tqdm